Moves HTTPClient into dedicated module
